refactor: extract funding RAP execution
